SpreadsheetFlexLayout: SpreadsheetDivComponent replaces DivElement
- SpreadsheetToolbarComponent.printTree previously seemed to be missing
